Remove _allocate() in favor of open()
Since the goal was to properly handle "out of disk space" error, allocate() only partially fixes the problem. Also allocate() introduces some new issues related to the disadvantages of sparse files. A better solution is to have a custom open() that knows how to properly handle said error. This open() is meant to be used by every method that needs to read/write a file.