Revert Dynaconf validation (commit fea38242)
Previously, Dynaconf validation was added to replace the manual validation which reduced the number of code lines. However, it seems once Dynaconf applies its validation for a certain key, it is no longer confident of the source configuration file from which key:value field is coming from, so it gives "undefined" for the field's "source" to inspect.get_history(). This breaks show-config's source indicator. Therefore, we are back to manual validation.