Remove support for passing post data as arguments
Temporarily deprecate support for passing post data as arguments. This was mainly because of strange Typer error "RuntimeError: Type not yet supported: <class 'typer.models.Context'>" when elapi is run on Python 3.9. Upon diagnosing, the culprit turned out to be the use of typer.Context which elapi cli post command was using. typer.Context works fine on Python 3.11.

Support for said feature should be brought back once Python 3.11 becomes the default on Debian.