Add support for format extension name convention
This is mainly to add proper support for "yaml" and "yml" extensions under YAMLFormat class. The "convention" attribute/property allows adding one or a list of extensions that the formatter class wants to support. If there aren't more than one convention then it would be name == convention. Name attribute should no longer be called to use as extension name!

Convention allows variance in extension names. It should be up to the client interface (in this case cli/elapi.py) which convention to prioritize and use. Ideally, the convention that is used the most should come first in the last. E.g.,
- convention for JSON: name = "json"; convention = name
- convention for TOML: name = "toml"; convention = name
- convention for YAML: name = "yaml"; convention = ["yaml", "yml"]
- convention for XSS: name = "xss"; convention = ["xss", "rss"]