Always validate configuration
Scripts or elAPI plugins, configuration file is always validated "only once", unless "developer_mode" is true. I.e., running reinitiate_config is no longer necessary to reinitiate configuration file for scripts using elAPI.
When developer_mode is on, validation is not run, and if a configuration key is missing, a PreventiveWarning exception is thrown with a clear message that explains why the error is necessary. I.e., elAPI knows missing with an absence of a configuration key and without validation, most HTTP APIs will fail.