Handle plugin config values passed to --OC
"plugins" in elapi.yml can contain a dictionary for each "plugin": "plugin name" as key, and "plugin configuration" as configuration value. "plugin configuration" must itself be a dictionary. This allows --OC to be able to update any value easily found in "plugin configuration". This commit adds the change. --OC also makes sure if a "plugin" under "plugins" is ignored by PluginConfigurationValidator, the same "plugin" passed to --OC is honored. An appropriate warning will also be shown to not confuse the user.